Deleted login component (because of login page) and renamed folder to authentication as this is what it does
