+++ sonarQuest-frontend/src/app/authentication/Token.ts
