+++ sonarQuest-frontend/src/app/authentication/authentication.guard.ts
