+++ sonarQuest-frontend/src/app/authentication/local-storage.service.ts
