+++ sonarQuest-frontend/src/app/authentication/username-password.ts
