--- sonarQuest-frontend/src/app/login/login.component.css
@@ -1,3 +1,0 @@
-.buttons {
-    float: right;
-}
