--- sonarQuest-frontend/src/app/login/login.component.html
@@ -1,32 +1,0 @@
-<mat-toolbar color="primary">
-  <h2 mat-dialog-title>{{"LOGIN.LOGIN"| translate}}</h2>
-  <span class="toolbar-spacer"></span>  
-  <button mat-button mat-raised-button color="" mat-dialog-close
-    [matTooltip]="'LOGIN.CANCEL' | translate"
-    [matTooltipClass]="'tooltipMultiline'"
-    matTooltipPosition="below">
-    <i class="ra ra-2x ra-crossed-bones"></i>
-  </button>
-</mat-toolbar>
-<mat-dialog-content>
-  <div layout="column" layout-padding>
-    <mat-form-field>
-      <input matInput placeholder='{{"LOGIN.USERNAME"| translate}}' [(ngModel)]="username"  (keydown)="onKeyDown($event)"  required >
-    </mat-form-field>
-
-    <mat-form-field>
-      <input matInput placeholder='{{"LOGIN.PASSWORD"| translate}}' [(ngModel)]="password"  (keydown)="onKeyDown($event)"  required [type]="'password'">
-    </mat-form-field>
-  </div>
-</mat-dialog-content>
-      
-<mat-dialog-actions>
-    <span class="toolbar-spacer"></span>
-     
-    <button mat-button mat-raised-button color="primary" type="submit" (click)="onSubmit()"
-      [matTooltip]="'LOGIN.LOGIN' | translate"
-      [matTooltipClass]="'tooltipMultiline'"
-      matTooltipPosition="below">
-      <i class="ra ra-4x ra-key"></i>
-    </button>
-</mat-dialog-actions>
