+++ sonarQuest-frontend/src/app/pages/login-page/login-page.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {AuthenticationService} from "../../login/authentication.service";
+import {AuthenticationService} from "../../authentication/authentication.service";
 
 @Component({
   selector: 'sq-login-page',
